Updated most things to produce a quality sine wave. Unfortunately still running into render issues on the first half cycle. Check outside of virtual box

+++ src/Source/Synth/GPCore.h
@@ -22,7 +22,7 @@ class GPInfo {
         int nextNetworkID;
 
         // special node values
-        float *time;
+        double *time;
         float *cps;
 
         // chance params
